fix(nextjs): refresh stale 31337 EFSBytesStore ABI (statusCode uint16)
deploy-pin-check flagged a drift: the committed 31337 block had
`request()` returning `uint256 statusCode`, but current contract source
(EFSBytesStore/EFSRouter, ADR-0058) returns `uint16` — so a fresh local
fork deploy regenerates `uint16` and the check failed.

This is a pre-existing drift: a prior Sepolia-only deploy refreshed the
11155111 block but, via the per-chain merge (ADR-0061), left 31337
stale. Regenerating on a fresh pinned fork brings 31337 current
(addresses unchanged — pin held; ABI-only, 2 lines). 5318008 is already
current; 11155111 is left as-is (it reflects the actual frozen Sepolia
deploy, which pin-check never regenerates).

Verified: fresh `yarn fork` + `yarn deploy` at FORK_BLOCK=10691000;
diff is statusCode uint256->uint16 only, no address changes.

Permanence-tier: Durable
Refs: ADR-0058, ADR-0061

+++ packages/nextjs/contracts/deployedContracts.ts
@@ -1462,9 +1462,9 @@ const deployedContracts = {
           name: "request",
           outputs: [
             {
-              internalType: "uint256",
+              internalType: "uint16",
               name: "statusCode",
-              type: "uint256",
+              type: "uint16",
             },
             {
               internalType: "bytes",
